refactor(decisions): extract shared mkdir-lock.cjs and remove dead code - Extract acquireMkdirLock/releaseLock into scripts/hooks/lib/mkdir-lock.cjs; both json-helper.cjs and render-decisions.cjs had identical copies - Remove the passthrough initDecisionsContent wrapper in json-helper.cjs; import directly from decisions-format.cjs as the canonical source - Fix dollar-EXIT -> dollar-underscore-EXIT variable reference bug in dream-commit debug log - Remove unused consumedObsIds Set in decisions-ledger-migration.ts
